requirements.txt
streamlit
openai-agents
tavily-python
python-dotenv
requests
better_profanity
pydantic